Refactor user id type handling for consistency
* web interface에서는 user id type을 String 타입으로 handling, 이외 내부에서는 Long 타입으로 handling 하도록 수정

+++ src/main/kotlin/com/slothhell/api/config/security/ServiceOAuth2AuthenticationSuccessHandler.kt
@@ -24,8 +24,8 @@ class ServiceOAuth2AuthenticationSuccessHandler(
 	) {
 		val userName = authentication.extractOAuth2UserName()
 
-		val accessToken: String = jwtAuthenticationProvider.generateAccessToken(userName)
-		val refreshToken: String = jwtAuthenticationProvider.generateRefreshToken(userName)
+		val accessToken: String = jwtAuthenticationProvider.generateAccessToken(userName.toLong())
+		val refreshToken: String = jwtAuthenticationProvider.generateRefreshToken(userName.toLong())
 
 		val responseBody = mapOf(
 			"accessToken" to accessToken,
